==> 04_Models/jupyter_notebooks/08_ORA_gmm.ipynb <==
### Set parameters and names

import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt

import gseapy as gp
from gseapy import barplot, dotplot
import pickle        


mpl.rc('axes', labelsize=14)
mpl.rc('xtick', labelsize=12)
mpl.rc('ytick', labelsize=12)
plt.rcParams['savefig.dpi'] = 250
plt.rcParams['figure.dpi'] = 100
plt.rcParams['axes.facecolor'] = 'white'
%matplotlib inline

INPUT_GENES    = 'ALL'
INPUT_FEATURES = 'X_FC'
INPUT_NORM     = 'z'
CODINGS_SIZE = 6

ID     = f'{CODINGS_SIZE}D_{INPUT_GENES}_{INPUT_FEATURES}_{INPUT_NORM}'
METHOD = 'VAE'
k      = 80
LABELS_COL = f'GMM_{METHOD}_{k}'
DIR_DATA= f'../data/{ID}_analysis/{LABELS_COL}/'
DIR_FIG = f'../figures/{ID}_analysis/{METHOD}/ORA_k{k}'
! mkdir -p {DIR_FIG}



with open(f'{DIR_DATA}gene_clusters_dict.pkl', 'rb') as f:
    GENE_CLUSTERS = pickle.load(f)



MOUSE_DB = gp.get_library_name(organism='Mouse')
MOUSE_DB


RESULTS = {}
for k, value in GENE_CLUSTERS.items():
    enr = gp.enrichr(gene_list= value['gene_list'],
                    background=  f'../data/{ID}_analysis/background.list',
                    gene_sets=['KEGG_2019_Mouse','WikiPathways_2019_Mouse','GO_Biological_Process_2023','GO_Molecular_Function_2023'],
                    organism='Mouse', 
                    outdir=None, 
                    verbose=True )
    
    RESULTS[f'Cluster_{k}']= {'result': enr.results, 'n_genes':value['len']}

### Remove from Term final string after '('

RESULTS_CLEANED = RESULTS.copy()
for key,value in RESULTS_CLEANED.items():
    x=RESULTS_CLEANED[key]['result']['Term'].str
    NEW_COL=[]
    for el in list(x.split('(')):
        NEW_COL.append(el[0])
    RESULTS_CLEANED[key]['result']['Term']=NEW_COL

COL_DICT = {
    'KEGG_2019_Mouse': '#66C2A5',
    'WikiPathways_2019_Mouse': '#FC8D62',
    'GO_Biological_Process_2023': '#8DA0CB',
    'GO_Molecular_Function_2023': '#E78AC3'
}


for cluster_k, value in RESULTS_CLEANED.items():
        n_genes=RESULTS_CLEANED[cluster_k]['n_genes']
        
        df = RESULTS_CLEANED[cluster_k]['result']
                #
        try:
                barplot(df,
                        column="Adjusted P-value",
                        group='Gene_set',
                        color=COL_DICT,
                        top_term=5,
                        title=f'{cluster_k}, n={n_genes}',
                        #xticklabels_rot=45,
                        #show_ring=False,
                        #marker='o',
                        #cmap='viridis'
                        )
                plt.savefig(f'{DIR_FIG}/{cluster_k}.pdf', format="pdf", bbox_inches="tight")
        except ValueError:
                with open(f'{DIR_FIG}/{cluster_k}_no_sign_terms.error', "w") as file:
                        file.write("No significant terms")                    
        if cluster_k=='5':break
# Adjust layout
plt.tight_layout()
# Show the plot
plt.show()
